Title: 바이러스 검사, Time: 366ms, Memory: 4MB, Status: Passed - CodeTree

+++ 240621/바이러스 검사/virus-detector.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int n;
 int restaurantCustomers[1000001];
 int coverage_jang, coverage_won;
-int minInspectors;
+long long int minInspectors;
 
 int main()
 {
